Added generation of clipPaths for layers and margin settings for isobands etc
If clipping is enabled using clip=true, most data dependent layers
will be preceded by a clipPath with a rectangle of the size of the
layer itself, and the clip-path attribute for the layer is set.

This is indended to be used with the margin settings to enable
isolines, maps, numbers, symbols etc to be generated for a slightly
larger area than reserved to the layer itself so that features
slightly outside the layer may still be partially visible even
if their reference point is outside the layer.

* lines may be several pixels wide
* symbols may be large
* numbers may be large
* arrows may be large
* visual SVG/CSS effects may extend the region covered by any object