Added GeoJSON support
- Setting the desired output type now automatically selects the default template
- Added geojson.tmpl
- Added GeoJSON formatting of isobands, isolines and maps

At the same time removed unnecessary SmartMet:: specifiers when
already in SmartMet namespace.

Unfortunately GDAL formats GeoJSON with full precision even though the
RFC mentions that with 6 decimals the accuracy is 10 cm, and that
WGS84 is off by a couple meters in many places. Should consider
writing an own formatter instead of taking the coordinates as a
substring from GDAL output.