Bug fix release: similarly named structs caused wrong linkage and crashes
ArrowLayer and SymbolLayer both define PointValue struct but with different
contents since arrows need both direction and speed. This causes random
behaviour after the the first element in a vector of PointValues has been
traversed, and eventually a segmentation fault. The structs are now placed
in an anonymous namespace to prevent collisions.