Merge scale value into transform attgributes when needed
librsvg2 will ignore the scale setting if the element is defined like this:

<use xlink:href="#symbol" scale="0.5" transform="translate(x y) rotate(alpha)"/>

The code will now automatically convert the output into form

<use xlink:href="#symbol" transform="translate(x y) rotate(alpha) scale(0.5)"/>

which librsvg2 handles correctly. Modified ArrowLayer, SymbolLayer and NumberLayer
accordingly. Note that if the layer defines an extra scale setting, it will be
multiplied with the SVG attribute value.