Added regex support for WMS GetCapabilities namespace queries
If the namespace parameter is surrounded by slashes, the parameter is
interpreted to be a regex which must match some part of the layer
title for the layer to be included into the GetCapabilities response.
Otherwise the string is expected to match the namespace of the title.
All comparisons are case insensitive.

Examples:
* namespace=fmi
* namespace=/hirlam|ecmwf/
* namespace=/temperature|wind/