Fixed code to allow both double and string fmisids (BRAINSTORM-1016)
The code now uses an utility function called get_fmisid to process
the variant object returned by observation-engine. The code now
allows the value of fmisid to be an integer, a double or a string.
The latter two will be converted to integers.

The engine should return only integers. Once the engine has been
fixed, the function should be modified to also allow only integers.