New release: added station layout support for symbol, number and arrow labels
In station-layout the positions are defined by an array of station
definitions declared with variable "stations". If the array element is
an integer, it is assumed to be an fmisid. Otherwise the element is
assumed to be an object which also declares at least of of fmisid,
wmo, lpnn, geoid or latitude and longitude.  The object may also
declare optional dx,dy adjustments for the data associated with the
station. The adjustments do not modify the position of the station,
for example any marker placed on the station will still be exactly at
the station coordinate.  The adjustments do apply to numbers, weather
symbols and wind arrows based on the observed values at the station.

The adjustments are multilevel as follows:

- for symbols and arrows one can define global dx and dy settings
- for numbers the dx and dy settings are defined in the labels section
- the individual dx dy adjustments defined for individual stations
  are added to the global adjustments

The interpretation of dx and dy when defining the Positions object is
now changed. If the dx and dy generate a grid, they define exact
coordinate adjustments and do not specify adjustments for associated
data. For latlon and station layouts the dx and dy variables do
specify adjustments for the associated data.