Layer specific precisions for SVG paths are now configurable BRAINSTORM-1155)
The default precision for all layers is 1.0 digits. The default or individial
prevision for a layer can be overridden by using the configuration file:

precision:
{
	default	= 0.3;
//	isoband	= 0.3;
//	isoline	= 0.3;
//	frame	= 0.3;
//	icemap	= 0.3;
//	map	= 0.3;
//	postgis	= 0.3;
//	wkt	= 0.3;
};

A default prevision of 0.3 means each number is rounded roughly to
half a pixel resolution, since 10**(-0.5) ~ 0.501. Even less resolution
such as 0.1 (10**(-0.1) ~ 0.79) might often be useful.

The numbers will be printed to a number of decimals which will cover
the intended precision. For example, for precisions 0.1 ... 1.0 one
decimal will be printed. However, since the rounding prevision varies,
and a sequence of identical coordinates will be reduced to one, the
number of printed vertices may be significantly lower with lower
precisions.

The precision can also be set in the JSON for each layer in case there
is a need to adjust the defaults.