Use only one TemplateFactory object, WMSGetCapabilities is now plain namespace instead of an object
Previously WMSGetCapabilities was a class, which initialized on
construction a TemplateFactory object, asked it for a template and
stored it for internal use. The idea is to use a single
TemplateFactory as provided by Plugin so that templates can be cached.
After refactoring WMSGetCapabilities no longer had a state, so it
was changed into a namespace with a single function.