Added unit conversions. Fixed arrow flopping, added flipping (BRAINSTORM-1515, BRAINSTORM-1516)
Previously one had to specify an explicit multiplier and/or offset to perform unit conversions.
Now the WMS configuration file can be used to list known conversions as follows:

unit_conversion:
{
        ms_to_knots:
        {
                multiplier      = 1.94384449244;
        },
        celsius_to_fahrenheit:
        {
                multiplier      = 1.8;
                offset          = 32;
        },
        celsius_to_kelvin:
        {
                offset          = 273.15;
        }
}

The conversion can then be used in IsobandLayer, IsolineLayer,
NumberLayer, ArrowLayer, Intersection and Label settings as follows:

  "unit_conversion": "ms_to_knots"

southflip is used to specify that arrows should be flopped in the
southern hemisphere. Typically this is used when rendering wind barbs,
whose barbs point to the low pressure center. If the wind barbs have
been drawn in the opposite direction, northflop can be set to flop
only in the northern hemisphere. The default is not to flop at all.

The equivalent flip setting will always cause arrows to be flipped.
This may be useful for example when reusing wind arrows for example
for flow arrows, since the numerical interpretation of wind directions
can change from "wind blows from" to "water/ice flows to".